feat(#37): rename packages
