+++ src/main/java/org/eolang/opeo/CompileMojo.java
@@ -28,7 +28,7 @@ import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
-import org.eolang.opeo.compiler.Compiler;
+import org.eolang.opeo.compilation.Compiler;
 
 /**
  * Compiles high-level EO representation into low-level representation.
