+++ src/main/java/org/eolang/opeo/DecompileMojo.java
@@ -28,7 +28,7 @@ import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
-import org.eolang.opeo.vmachine.Decompiler;
+import org.eolang.opeo.decompilation.Decompiler;
 
 /**
  * Decompiles bytecode in EO representation into high-level EO representation.
