+++ src/main/java/org/eolang/opeo/decompilation/LocalVariables.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.opeo.vmachine;
+package org.eolang.opeo.decompilation;
 
 import java.util.ArrayList;
 import java.util.HashMap;
