+++ src/main/java/org/eolang/opeo/jeo/JeoDecompiler.java
@@ -30,8 +30,8 @@ import java.util.stream.Collectors;
 import org.eolang.jeo.representation.xmir.XmlMethod;
 import org.eolang.jeo.representation.xmir.XmlNode;
 import org.eolang.jeo.representation.xmir.XmlProgram;
-import org.eolang.opeo.vmachine.DecompilerMachine;
-import org.eolang.opeo.vmachine.LocalVariables;
+import org.eolang.opeo.decompilation.DecompilerMachine;
+import org.eolang.opeo.decompilation.LocalVariables;
 import org.w3c.dom.Node;
 import org.xembly.Transformers;
 import org.xembly.Xembler;
